Update the pre-launch content to match the mocks, and add a favicon
