+++ frontend/package.json
@@ -48,10 +48,8 @@
     "eject": "react-scripts eject",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public",
-
     "start:development": "env-cmd -f .env.development npm run start",
     "storybook:development": "env-cmd -f .env.development npm run storybook",
-
     "build:staging": "env-cmd -f .env.staging npm run build",
     "build:prod": "env-cmd -f .env.prod npm run build",
     "build:development": "env-cmd -f .env.development npm run build",
