+++ frontend/src/pages/Landing/PreLaunchSiteContent.module.scss
@@ -1,0 +1,39 @@
+@import "../../styles/variables";
+
+// TODO: merge these styles into MaterialTheme.tsx, App.module.scss, and
+// LandingPage.module.scss, and delete this file.
+.App {
+  text-align: center;
+  position: relative;
+  min-height: 100%;
+  background-color: #4733bf;
+}
+
+.Content {
+  padding-bottom: $footer-size;
+  height: 100%;
+}
+
+.PreLaunchSnippet {
+  color: #ffffff;
+  text-align: left;
+  padding: 28px;
+}
+
+.Grid {
+  margin: auto;
+  max-width: $max-width;
+}
+
+.HomeLogo {
+  flex-grow: 1;
+  text-align: left;
+  font-family: "Montserrat" !important;
+  color: #4733bf;
+  padding-left: 20px;
+  a:link,
+  a:visited {
+    color: $white;
+    text-decoration: none;
+  }
+}
