+++ frontend/src/utils/AppBanner.module.scss
@@ -1,0 +1,7 @@
+@import "../styles/variables";
+
+.Banner {
+  background-color: #000000;
+  color: #ffffff;
+  padding: 7px;
+}
